split code from storageunit to separate files
