+++ storageUnit/storageunit_bench_test.go
@@ -1,0 +1,83 @@
+package storageUnit_test
+
+import (
+	"fmt"
+	"math/rand"
+	"strconv"
+	"testing"
+)
+
+const (
+	valuesInDb = 100000
+)
+
+func logError(err error) {
+	if err != nil {
+		fmt.Println(err.Error())
+	}
+}
+
+func BenchmarkStorageUnit_Put(b *testing.B) {
+	b.StopTimer()
+	s := initStorageUnit(b, 1)
+	defer func() {
+		err := s.DestroyUnit()
+		logError(err)
+	}()
+	b.StartTimer()
+
+	for i := 0; i < b.N; i++ {
+		b.StopTimer()
+		nr := rand.Intn(valuesInDb)
+		b.StartTimer()
+
+		err := s.Put([]byte(strconv.Itoa(nr)), []byte(strconv.Itoa(nr)))
+		logError(err)
+	}
+}
+
+func BenchmarkStorageUnit_GetWithDataBeingPresent(b *testing.B) {
+	b.StopTimer()
+	s := initStorageUnit(b, 1)
+	defer func() {
+		err := s.DestroyUnit()
+		logError(err)
+	}()
+	for i := 0; i < valuesInDb; i++ {
+		err := s.Put([]byte(strconv.Itoa(i)), []byte(strconv.Itoa(i)))
+		logError(err)
+	}
+	b.StartTimer()
+
+	for i := 0; i < b.N; i++ {
+		b.StopTimer()
+		nr := rand.Intn(valuesInDb)
+		b.StartTimer()
+
+		_, err := s.Get([]byte(strconv.Itoa(nr)))
+		logError(err)
+	}
+}
+
+func BenchmarkStorageUnit_GetWithDataNotBeingPresent(b *testing.B) {
+	b.StopTimer()
+	s := initStorageUnit(b, 1)
+	defer func() {
+		err := s.DestroyUnit()
+		logError(err)
+	}()
+	for i := 0; i < valuesInDb; i++ {
+		err := s.Put([]byte(strconv.Itoa(i)), []byte(strconv.Itoa(i)))
+		logError(err)
+	}
+	b.StartTimer()
+
+	for i := 0; i < b.N; i++ {
+		b.StopTimer()
+		nr := rand.Intn(valuesInDb) + valuesInDb
+		b.StartTimer()
+
+		_, err := s.Get([]byte(strconv.Itoa(nr)))
+		logError(err)
+	}
+}
